Mask password characters
This commit replaces the inserted plain characters in the input panel displayed with star characters, i.e. it effectively hides the password during the interaction with the user.

Still, the commit needs improvements. For now, the only way to hide the password is to hide the input panel on each character entered/deleted (`on_change` method) and show a new input panel with star characters displayed. Hiding input panel invokes `on_cancel` method - which is bad, since ESC key also invokes the same method. it is not possible to make a separation between closing the input panel using ESC key and the logic currently implemented in `on_cancel` method. There is a way to do it using sublime's `on_query_context` method, but that requires a significant effort. Currently, it is possible to close the input panel by deleting all the characters inserted.